+++ jdaviz/configs/specviz2d/plugins/__init__.py
@@ -1,1 +1,2 @@
 from .spectral_extraction.spectral_extraction import * # noqa
+from .cross_dispersion_profile.cross_dispersion_profile import * # noqa